testing/requirements.txt
# these required on only for testing purposes
pytest
zbar-py
Pillow
